[REFACTOR] useModal hook 분리
